Add files via upload
"LinkedIn Data Professional Salary and Location Satisfaction Dashboard"
The dashboard is centered around survey data collected from LinkedIn users, specifically focusing on their satisfaction with salary and location. It also mentions that the dashboard is based on raw data analysis, which provides context to the audience about the nature of the content.